refactor: move each class to its package to organize the structure
