+++ src/main/kotlin/racingcar/car/Car.kt
@@ -1,4 +1,6 @@
-package racingcar
+package racingcar.car
+
+import racingcar.strategy.MoveStrategy
 
 data class Car(
     val name: String,
