+++ src/main/kotlin/racingcar/strategy/MoveStrategy.kt
@@ -1,0 +1,7 @@
+package racingcar.strategy
+
+import camp.nextstep.edu.missionutils.Randoms
+
+fun interface MoveStrategy {
+    fun canMove(): Boolean
+}
