+++ src/test/kotlin/racingcar/car/CarTest.kt
@@ -1,9 +1,10 @@
-package racingcar
+package racingcar.car
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
+import racingcar.strategy.MoveStrategy
 
 class CarTest {
